+++ src/vp8/common/filter.rs
@@ -3,103 +3,103 @@ pub const VP8_FILTER_SHIFT: i32 = 7 as i32;
 #[unsafe(no_mangle)]
 pub static mut vp8_bilinear_filters: [[i16; 2]; 8] = [
     [
-        128 as i32 as i16,
-        0 as i32 as i16,
+        128 as i16,
+        0 as i16,
     ],
     [
-        112 as i32 as i16,
-        16 as i32 as i16,
+        112 as i16,
+        16 as i16,
     ],
     [
-        96 as i32 as i16,
-        32 as i32 as i16,
+        96 as i16,
+        32 as i16,
     ],
     [
-        80 as i32 as i16,
-        48 as i32 as i16,
+        80 as i16,
+        48 as i16,
     ],
     [
-        64 as i32 as i16,
-        64 as i32 as i16,
+        64 as i16,
+        64 as i16,
     ],
     [
-        48 as i32 as i16,
-        80 as i32 as i16,
+        48 as i16,
+        80 as i16,
     ],
     [
-        32 as i32 as i16,
-        96 as i32 as i16,
+        32 as i16,
+        96 as i16,
     ],
     [
-        16 as i32 as i16,
-        112 as i32 as i16,
+        16 as i16,
+        112 as i16,
     ],
 ];
 #[unsafe(no_mangle)]
 pub static mut vp8_sub_pel_filters: [[i16; 6]; 8] = [
     [
-        0 as i32 as i16,
-        0 as i32 as i16,
-        128 as i32 as i16,
-        0 as i32 as i16,
-        0 as i32 as i16,
-        0 as i32 as i16,
+        0 as i16,
+        0 as i16,
+        128 as i16,
+        0 as i16,
+        0 as i16,
+        0 as i16,
     ],
     [
-        0 as i32 as i16,
+        0 as i16,
         -(6 as i32) as i16,
-        123 as i32 as i16,
-        12 as i32 as i16,
+        123 as i16,
+        12 as i16,
         -(1 as i32) as i16,
-        0 as i32 as i16,
+        0 as i16,
     ],
     [
-        2 as i32 as i16,
+        2 as i16,
         -(11 as i32) as i16,
-        108 as i32 as i16,
-        36 as i32 as i16,
+        108 as i16,
+        36 as i16,
         -(8 as i32) as i16,
-        1 as i32 as i16,
+        1 as i16,
     ],
     [
-        0 as i32 as i16,
+        0 as i16,
         -(9 as i32) as i16,
-        93 as i32 as i16,
-        50 as i32 as i16,
+        93 as i16,
+        50 as i16,
         -(6 as i32) as i16,
-        0 as i32 as i16,
+        0 as i16,
     ],
     [
-        3 as i32 as i16,
+        3 as i16,
         -(16 as i32) as i16,
-        77 as i32 as i16,
-        77 as i32 as i16,
+        77 as i16,
+        77 as i16,
         -(16 as i32) as i16,
-        3 as i32 as i16,
+        3 as i16,
     ],
     [
-        0 as i32 as i16,
+        0 as i16,
         -(6 as i32) as i16,
-        50 as i32 as i16,
-        93 as i32 as i16,
+        50 as i16,
+        93 as i16,
         -(9 as i32) as i16,
-        0 as i32 as i16,
+        0 as i16,
     ],
     [
-        1 as i32 as i16,
+        1 as i16,
         -(8 as i32) as i16,
-        36 as i32 as i16,
-        108 as i32 as i16,
+        36 as i16,
+        108 as i16,
         -(11 as i32) as i16,
-        2 as i32 as i16,
+        2 as i16,
     ],
     [
-        0 as i32 as i16,
+        0 as i16,
         -(1 as i32) as i16,
-        12 as i32 as i16,
-        123 as i32 as i16,
+        12 as i16,
+        123 as i16,
         -(6 as i32) as i16,
-        0 as i32 as i16,
+        0 as i16,
     ],
 ];
 unsafe fn filter_block2d_first_pass(
@@ -122,25 +122,25 @@ unsafe fn filter_block2d_first_pass(
                 Temp = *src_ptr.offset(
                     (-(2 as i32) * pixel_step as i32) as isize,
                 ) as i32
-                    * *vp8_filter.offset(0 as i32 as isize) as i32
+                    * *vp8_filter.offset(0 as isize) as i32
                     + *src_ptr.offset(
                         (-(1 as i32) * pixel_step as i32) as isize,
                     ) as i32
-                        * *vp8_filter.offset(1 as i32 as isize)
+                        * *vp8_filter.offset(1 as isize)
                             as i32
-                    + *src_ptr.offset(0 as i32 as isize) as i32
-                        * *vp8_filter.offset(2 as i32 as isize)
+                    + *src_ptr.offset(0 as isize) as i32
+                        * *vp8_filter.offset(2 as isize)
                             as i32
                     + *src_ptr.offset(pixel_step as isize) as i32
-                        * *vp8_filter.offset(3 as i32 as isize)
+                        * *vp8_filter.offset(3 as isize)
                             as i32
                     + *src_ptr.offset((2 as u32).wrapping_mul(pixel_step) as isize)
                         as i32
-                        * *vp8_filter.offset(4 as i32 as isize)
+                        * *vp8_filter.offset(4 as isize)
                             as i32
                     + *src_ptr.offset((3 as u32).wrapping_mul(pixel_step) as isize)
                         as i32
-                        * *vp8_filter.offset(5 as i32 as isize)
+                        * *vp8_filter.offset(5 as isize)
                             as i32
                     + (VP8_FILTER_WEIGHT >> 1 as i32);
                 Temp >>= VP8_FILTER_SHIFT;
@@ -179,23 +179,23 @@ unsafe fn filter_block2d_second_pass(
             while j < output_width {
                 Temp = *src_ptr.offset(
                     (-(2 as i32) * pixel_step as i32) as isize,
-                ) * *vp8_filter.offset(0 as i32 as isize)
+                ) * *vp8_filter.offset(0 as isize)
                     as i32
                     + *src_ptr.offset(
                         (-(1 as i32) * pixel_step as i32) as isize,
-                    ) * *vp8_filter.offset(1 as i32 as isize)
+                    ) * *vp8_filter.offset(1 as isize)
                         as i32
-                    + *src_ptr.offset(0 as i32 as isize)
-                        * *vp8_filter.offset(2 as i32 as isize)
+                    + *src_ptr.offset(0 as isize)
+                        * *vp8_filter.offset(2 as isize)
                             as i32
                     + *src_ptr.offset(pixel_step as isize)
-                        * *vp8_filter.offset(3 as i32 as isize)
+                        * *vp8_filter.offset(3 as isize)
                             as i32
                     + *src_ptr.offset((2 as u32).wrapping_mul(pixel_step) as isize)
-                        * *vp8_filter.offset(4 as i32 as isize)
+                        * *vp8_filter.offset(4 as isize)
                             as i32
                     + *src_ptr.offset((3 as u32).wrapping_mul(pixel_step) as isize)
-                        * *vp8_filter.offset(5 as i32 as isize)
+                        * *vp8_filter.offset(5 as isize)
                             as i32
                     + (VP8_FILTER_WEIGHT >> 1 as i32);
                 Temp >>= VP8_FILTER_SHIFT;
@@ -235,7 +235,7 @@ unsafe fn filter_block2d(
             HFilter,
         );
         filter_block2d_second_pass(
-            (&raw mut FData as *mut i32).offset(8 as i32 as isize),
+            (&raw mut FData as *mut i32).offset(8 as isize),
             output_ptr,
             output_pitch,
             4 as u32,
@@ -299,7 +299,7 @@ pub unsafe fn vp8_sixtap_predict8x8_c(
             HFilter,
         );
         filter_block2d_second_pass(
-            (&raw mut FData as *mut i32).offset(16 as i32 as isize),
+            (&raw mut FData as *mut i32).offset(16 as isize),
             dst_ptr,
             dst_pitch,
             8 as u32,
@@ -337,7 +337,7 @@ pub unsafe fn vp8_sixtap_predict8x4_c(
             HFilter,
         );
         filter_block2d_second_pass(
-            (&raw mut FData as *mut i32).offset(16 as i32 as isize),
+            (&raw mut FData as *mut i32).offset(16 as isize),
             dst_ptr,
             dst_pitch,
             8 as u32,
@@ -375,7 +375,7 @@ pub unsafe fn vp8_sixtap_predict16x16_c(
             HFilter,
         );
         filter_block2d_second_pass(
-            (&raw mut FData as *mut i32).offset(32 as i32 as isize),
+            (&raw mut FData as *mut i32).offset(32 as isize),
             dst_ptr,
             dst_pitch,
             16 as u32,
@@ -401,11 +401,11 @@ unsafe fn filter_block2d_bil_first_pass(
         while i < height {
             j = 0 as u32;
             while j < width {
-                *dst_ptr.offset(j as isize) = ((*src_ptr.offset(0 as i32 as isize)
+                *dst_ptr.offset(j as isize) = ((*src_ptr.offset(0 as isize)
                     as i32
-                    * *vp8_filter.offset(0 as i32 as isize) as i32
-                    + *src_ptr.offset(1 as i32 as isize) as i32
-                        * *vp8_filter.offset(1 as i32 as isize)
+                    * *vp8_filter.offset(0 as isize) as i32
+                    + *src_ptr.offset(1 as isize) as i32
+                        * *vp8_filter.offset(1 as isize)
                             as i32
                     + VP8_FILTER_WEIGHT / 2 as i32)
                     >> VP8_FILTER_SHIFT)
@@ -435,14 +435,14 @@ unsafe fn filter_block2d_bil_second_pass(
         while i < height {
             j = 0 as u32;
             while j < width {
-                Temp = *src_ptr.offset(0 as i32 as isize) as i32
-                    * *vp8_filter.offset(0 as i32 as isize) as i32
+                Temp = *src_ptr.offset(0 as isize) as i32
+                    * *vp8_filter.offset(0 as isize) as i32
                     + *src_ptr.offset(width as isize) as i32
-                        * *vp8_filter.offset(1 as i32 as isize)
+                        * *vp8_filter.offset(1 as isize)
                             as i32
                     + VP8_FILTER_WEIGHT / 2 as i32;
                 *dst_ptr.offset(j as isize) =
-                    (Temp >> VP8_FILTER_SHIFT) as u32 as u8;
+                    (Temp >> VP8_FILTER_SHIFT) as u8;
                 src_ptr = src_ptr.offset(1);
                 j = j.wrapping_add(1);
             }
